Move s3 server mock to service package
