+++ packages/component-service-s3/mock/index.js
